Test file reorg
